+++ src/njoy21/input/PLOTR/Card6/Ystep/test/Ystep.test.cpp
@@ -11,7 +11,7 @@ using namespace njoy::njoy21::input;
 SCENARIO( "Ystep output values", "[PLOTR],[Card6], [Ystep]"){
 
   GIVEN( "valid Ystep parameters" ){
-    std::vector<float> validValues{0.1,0.5,0.8,1.0,2.0,80.0};
+    std::vector<double> validValues{0.1,0.5,0.8,1.0,2.0,80.0};
 
     THEN( "the returned class has the correct value" ){
       for( auto ystep : validValues ){
@@ -34,7 +34,7 @@ SCENARIO( "Ystep output values", "[PLOTR],[Card6], [Ystep]"){
   }//GIVEN
 
   GIVEN( "invalid Ystep parameters" ){
-    std::vector<float> invalidValues{ -2.0, -1.1, 0.0 };
+    std::vector<double> invalidValues{ -2.0, -1.1, 0.0 };
 
     THEN( "the class throws an exception" ){
       for( auto ystep : invalidValues ){
